feat(bin): prefer openssl@1.1
PHP does not compile on openssl@3. The "brew --prefix openssl" command
does not behave as expected. If openssl@1.1 is installed then it will
not output anything. If opsenssl@3 is installed then it will output a
prefix for openssl@3. This commit uses explicit 1.1 version of OpenSSL.

OpenSSL 3 was released to homebrew recently, so users may encounter this
issue after upgrading to newest versions.